fix: resolve MonoGame graphics build errors - fix uint to int conversion in WavFormatConverter - fix Color constructor ambiguity in RoomMeshRenderer (use byte instead of int) - fix int to uint conversions in MetalBackend - fix Vector3 array type mismatch in LightProbeSystem - fix ambiguous Color/Vector2 references in KotorGuiManager (use aliases) - convert Vector2 types for CalculateTextPosition method calls

+++ src/Andastra/Runtime/Graphics/MonoGame/Backends/MetalBackend.cs
@@ -956,9 +956,9 @@ namespace Andastra.Runtime.MonoGame.Backends
             {
                 InputElement element = layout.Elements[i];
                 MetalNative.SetVertexDescriptorAttribute(vertexDescriptor, (uint)i,
-                    ConvertTextureFormat(element.Format), element.Slot, element.AlignedByteOffset);
+                    ConvertTextureFormat(element.Format), (uint)element.Slot, (uint)element.AlignedByteOffset);
                 MetalNative.SetVertexDescriptorLayout(vertexDescriptor, (uint)element.Slot,
-                    element.AlignedByteOffset, 0); // stride would be calculated
+                    (uint)element.AlignedByteOffset, 0U); // stride would be calculated
             }
 
             return vertexDescriptor;
